+++ asf_search/Products/__init__.py
@@ -13,3 +13,4 @@ from .OPERAS1Product import OPERAS1Product  # noqa: F401
 from .ARIAS1GUNWProduct import ARIAS1GUNWProduct  # noqa: F401
 from .NISARProduct import NISARProduct  # noqa: F401
 from .ALOS2Product import ALOS2Product  # noqa: F401
+from .TROPOProduct import TROPOProduct # noqa: F401